refactor: move utils into utils2
