+++ twotone/tools/subtitles_fixer.py
@@ -8,53 +8,52 @@ from overrides import override
 from tqdm import tqdm
 from tqdm.contrib.logging import logging_redirect_tqdm
 
-from . import utils
 from .tool import Tool
-from .utils2 import generic, process
+from twotone.tools.utils2 import generic, process, subtitles as subs, video as vid
 
 
-class Fixer(utils.InterruptibleProcess):
+class Fixer(generic.InterruptibleProcess):
     def __init__(self, logger: logging.Logger, really_fix: bool):
         super().__init__()
         self.logger = logger
         self._do_fix = really_fix
 
-    def _print_broken_videos(self, broken_videos_info: [(utils.VideoInfo, [int])]):
+    def _print_broken_videos(self, broken_videos_info: [(vid.VideoInfo, [int])]):
         self.logger.info(f"Found {len(broken_videos_info)} broken videos:")
         for broken_video in broken_videos_info:
             self.logger.info(f"{len(broken_video[1])} broken subtitle(s) in {broken_video[0].path} found")
 
-    def _no_resolver(self, video_track: utils.VideoTrack, content: str):
+    def _no_resolver(self, video_track: vid.VideoTrack, content: str):
         self.logger.error("Cannot fix the file, no idea how to do it.")
         return None
 
-    def _long_tail_resolver(self, video_track: utils.VideoTrack, content: str):
-        timestamps = list(utils.subrip_time_pattern.finditer(content))
+    def _long_tail_resolver(self, video_track: vid.VideoTrack, content: str):
+        timestamps = list(subs.subrip_time_pattern.finditer(content))
         last_timestamp = timestamps[-1]
-        time_from, time_to = map(utils.time_to_ms, last_timestamp.groups())
+        time_from, time_to = map(generic.time_to_ms, last_timestamp.groups())
         length = video_track.length
         new_time_to = min(time_from + 5000, length)
 
         begin_pos = last_timestamp.start(1)
         end_pos = last_timestamp.end(2)
 
-        content = content[:begin_pos] + f"{utils.ms_to_time(time_from)} --> {utils.ms_to_time(new_time_to)}" + content[end_pos:]
+        content = content[:begin_pos] + f"{generic.ms_to_time(time_from)} --> {generic.ms_to_time(new_time_to)}" + content[end_pos:]
         return content
 
-    def _fps_scale_resolver(self, video_track: utils.VideoTrack, content: str):
-        target_fps = utils.fps_str_to_float(video_track.fps)
-        multiplier = utils.ffmpeg_default_fps / target_fps
+    def _fps_scale_resolver(self, video_track: vid.VideoTrack, content: str):
+        target_fps = generic.fps_str_to_float(video_track.fps)
+        multiplier = subs.ffmpeg_default_fps / target_fps
 
-        return utils.alter_subrip_subtitles_times(content, multiplier)
+        return subs.alter_subrip_subtitles_times(content, multiplier)
 
     def _get_resolver(self, content: str, video_length: int):
-        timestamps = list(utils.subrip_time_pattern.finditer(content))
+        timestamps = list(subs.subrip_time_pattern.finditer(content))
         if len(timestamps) == 0:
             return self._no_resolver
 
         # check if last subtitle is beyond limit
         last_timestamp = timestamps[-1]
-        time_from, time_to = map(utils.time_to_ms, last_timestamp.groups())
+        time_from, time_to = map(generic.time_to_ms, last_timestamp.groups())
 
         if time_from < video_length and time_to > video_length:
             return self._long_tail_resolver
@@ -64,7 +63,7 @@ class Fixer(utils.InterruptibleProcess):
 
         return self._no_resolver
 
-    def _fix_subtitle(self, broken_subtitle, video_info: utils.VideoInfo) -> bool:
+    def _fix_subtitle(self, broken_subtitle, video_info: vid.VideoInfo) -> bool:
         video_track = video_info.video_tracks[0]
 
         with open(broken_subtitle, 'r', encoding='utf-8') as file:
@@ -82,13 +81,13 @@ class Fixer(utils.InterruptibleProcess):
                 file.write(new_content)
             return True
 
-    def _extract_all_subtitles(self,video_file: str, subtitles: [utils.Subtitle], wd: str) -> [utils.SubtitleFile]:
+    def _extract_all_subtitles(self,video_file: str, subtitles: [subs.Subtitle], wd: str) -> [subs.SubtitleFile]:
         result = []
         options = ["tracks", video_file]
 
         for subtitle in subtitles:
             outputfile = f"{wd}/{subtitle.tid}.srt"
-            subtitleFile = utils.SubtitleFile(path=outputfile, language=subtitle.language, encoding="utf8")
+            subtitleFile = subs.SubtitleFile(path=outputfile, language=subtitle.language, encoding="utf8")
 
             result.append(subtitleFile)
             options.append(f"{subtitle.tid}:{outputfile}")
@@ -97,7 +96,7 @@ class Fixer(utils.InterruptibleProcess):
 
         return result
 
-    def _repair_videos(self, broken_videos_info: [(utils.VideoInfo, [int])]):
+    def _repair_videos(self, broken_videos_info: [(vid.VideoInfo, [int])]):
         self._print_broken_videos(broken_videos_info)
         self.logger.info("Fixing videos")
 
@@ -127,7 +126,7 @@ class Fixer(utils.InterruptibleProcess):
                             # add fixed subtitles to video
                             self.logger.debug("Adding fixed subtitles to file")
                             temporaryVideoPath = video_file + ".fixed.mkv"
-                            utils.generate_mkv(input_video=video_without_subtitles, output_path=temporaryVideoPath, subtitles=subtitles)
+                            vid.generate_mkv(input_video=video_without_subtitles, output_path=temporaryVideoPath, subtitles=subtitles)
 
                             # overwrite broken video with fixed one
                             os.replace(temporaryVideoPath, video_file)
@@ -139,13 +138,13 @@ class Fixer(utils.InterruptibleProcess):
                     else:
                         self.logger.debug("Skipping video due to errors")
 
-    def _check_if_broken(self, video_file: str): # -> (utils.VideoInfo, [int]) | None:    // FIXME
+    def _check_if_broken(self, video_file: str): # -> (video.VideoInfo, [int]) | None:    // FIXME
         self.logger.debug(f"Processing file {video_file}")
 
         def diff(a, b):
             return abs(a - b) / max(a, b)
 
-        video_info = utils.get_video_data(video_file)
+        video_info = vid.get_video_data(video_file)
         video_length = video_info.video_tracks[0].length
 
         if video_length is None:
@@ -182,7 +181,7 @@ class Fixer(utils.InterruptibleProcess):
                 self._check_for_stop()
                 file_path = os.path.join(cd, file)
 
-                if utils.is_video(file_path):
+                if vid.is_video(file_path):
                     video_files.append(file_path)
 
         self.logger.debug("Analysing videos")
